+++ app/Http/Controllers/API/IdiomaController.php
@@ -8,6 +8,7 @@ use App\Models\Idiomas;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controllers\HasMiddleware;
 use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Gate;
 
 class IdiomaController extends Controller implements HasMiddleware
 {
@@ -39,6 +40,8 @@ class IdiomaController extends Controller implements HasMiddleware
      */
     public function store(Request $request)
     {
+        Gate::authorize('store', Idiomas::class);
+
         $idioma = json_decode($request->getContent(), true);
 
         $idioma = Idiomas::create($idioma);
@@ -60,6 +63,8 @@ class IdiomaController extends Controller implements HasMiddleware
      */
     public function update(Request $request, Idiomas $idioma)
     {
+        Gate::authorize('update', Idiomas::class);
+
         $cicloData = json_decode($request->getContent(), true);
         $idioma->update($cicloData);
 
@@ -71,6 +76,8 @@ class IdiomaController extends Controller implements HasMiddleware
      */
     public function destroy(Idiomas $idioma)
     {
+        Gate::authorize('destroy', Idiomas::class);
+
         try {
             $idioma->delete();
             return response()->json(null, 204);
